Add files via upload
This is a vacuum cleaner with a 10 X 10 matrix. The environment contains dirt and there are two agents which are used to clean the environment. One is table driven which works according the pattern of directions and the other one is simple reflex agent which moves in right direction by default and cleans the dirt.